+++ R/print.R
@@ -20,7 +20,7 @@
 print.lrtest <- function(x, ...) {
   print(paste("Log Likelihood Statistic:", round(x$statistic, 2), sep = " "))
   print(paste("p value:", round(x$p.value, 3), sep = " "))
-  if (class(x)[1] %in% c("one_sample_case_one", "one_sample_case_two")) {
+  if (class(x)[1] %in% c("one_sample_case_one", "one_sample_case_two", "one_sample_case_three", "one_sample_case_four")) {
     print(paste("Confidence Level: ", round(x[["conf.level"]][1], 3) * 100, "%", sep = ""))
     print(paste("Confidence Interval: (", round(x[["conf.int"]][1], 3), ", ", round(x[["conf.int"]][2], 3), ")", sep = ""))
   } else {